+++ CHANGELOG.rst
@@ -7,6 +7,10 @@ v31.0.0 (next)
 - WARNING: Drop support for Python 3.6 and 3.7. Add support for Python 3.10.
   Upgrade Django to version 4.x series.
 
+- Upgrade ScanCode-toolkit to version v31.
+  See https://github.com/nexB/scancode-toolkit/blob/develop/CHANGELOG.rst for an
+  overview of the changes in v31 compared to v30.
+
 - Implement run status auto-refresh using the htmx JavaScript library.
   The statuses of queued and running pipeline are now automatically refreshed
   in the project list and project details views every 10 seconds.
@@ -25,6 +29,9 @@ v31.0.0 (next)
 
 - Add new SCANCODEIO_REDIS_PASSWORD environment variable and setting
   to optionally set Redis instance password
+
+- Ensure a project cannot be deleted through the API while a pipeline is running.
+  https://github.com/nexB/scancode.io/issues/402
 
 v30.2.0 (2021-12-17)
 --------------------